Added New Version (See bottom of JS file)
I'm working on a new version that handles craft checking in case astronauts are in different places. There appears to be an issue with PebbleJS that prevents it from showing multiple notifications, but that isn't a problem. The problems lies in the fact that it will not even allow you to have multiple Pebble.showSimpleNotificationOnPebble() statements, even if all but one are unreachable. Any help is appreciated, but the new version should be up and running by next week.